Formatter: Add minimalistic support for moved file mapping.
Requires implementation in the UI.

The idea is that when a file has moved, Formatter will emit the
'missing-uri' signal.
Within that callback, the application should:
* ask the user for the new location
* call Formatter.addMapping()

Hopefully... if files have moved for a project, they most likely
have all moved in the same fashion, so this signal should only be
emitted once.