Refactor source factories a bit.
This change makes SourceFactory derived factories have an URI. This simplifies
the code that deals with source factories, as it can assume factory.uri to be
always there.
It also removes ObjectFactory.displayname, changing it to ObjectFactory.name
(ObjectFactory.name was previously either unused or used to store an URI...).
I also removed some "Just In Case" code (try/except: and the like) from
PictureFileSourceFactory._releaseBin.