Refactor interpolator code slightly.
Make track_object.interpolators a dictionary of the form property_name ->
(property, interpolator) so that we can lookup interpolators by property_name
without having to iterate all the keys every time.
Make it possible to rebuild interpolators when the content of a track_object's
bin changes.  This commit is part of a larger change to make interpolator operations undoable.