Make TrackObjects reusable once they have been removed from a Track.
This change makes TrackObjects reusable by decoupling construction and
initialization. Now you can call makeBin() and releaseBin() in pairs on the same
object to create/destroy bins keeping the same track object instance (with the
associated data, eg keyframes).
With this change, bins are now created when objects are added to a Track
(Track.addTrackObject) and destroyed when removed from a track
(Track.removeTrackObject)
The first visible benefit of this change is that it simplifies TimelineObject
do/undo operations a lot.