exportsettingswidget.py: remember window size closes bug 579237
This commit is a little complicated because the glade dialogs seemed to be
ignoring set_default_size. We connect to map event and use that to install
an idle handler which actually sets the dialog size. This seems to be the only
reliable way to set the size of this dialog.

The idle handler also connects to the configure event -- we don't want to do this
any earlier or the saved dialog size will be overwritten before the dialog is
shown.