medialibrary: Wait for all TrackObject-s to be in Tracks before seeking
We currently need to work around the fact that TrackObject-s are created
async and first added to the TimelineObject-s before the tracks. So we
need to wait the TrackObject to be in the tracks before the timeline duration
is updated, and the composition actually contains the gnlobject for the clips we
are adding to the timeline.