Prevent resizing the preferences dialog
Years down the road, it turns out that we still don't have enough
prefs to warrant scrollable areas in the preferences.

This commit doesn't remove the dimensions saving/loading code
but it prevents the window from resizing, effectively making it use
the smallest size required by its child widgets. If needed, this can
easily be reverted.