Allow entering a frame number into the time widget
The downside to this is that we don't seek live while typing anymore:
the frame number being typed would always get overridden by timecodes.
Users have to activate the text widget (press Enter) to seek with it.

On the other hand, this is compensated by the fact that we also seek
when the user defocuses the text widget.

Fixes bug #684236