Prevent the rendering progress dialog window from being resizable
This was the cause for the big empty space between the progress bar and
the action area: GTK3's size calculations were based on the height of the window
before its width would be resized.

Somehow, disabling window resizing works around this.

With the new text paragraph (or, alternatively, the default width of 400 px),
the dialog doesn't really need to be resized by users anyway.