For non-toolbar dialog buttons with labels, remove superfluous icons
The recent slew of changes in master added extra icons to various buttons.

In 2011 I would not have imagined that I'd be in favor of this in 2014,
but I've gotten used to not having random icons on buttons anymore.
It's also easier to support as an app maintainer, with more labeling freedom.

This commit also clarifies some button labels, replacing "Clear" by "Reset all".