viewer: Do not call the expose function when we set the aspect ratio
During caps renegotiation resulting from the change of the video
height/width, aspect ratio changes, there is a race in glimagesink where
we end up with the viewer displaying broken frames calling the expose()
method of the sink at the same time.

Avoid calling the videosink.expose() while we change the resolution
in the pipeline.

https://bugzilla.gnome.org/show_bug.cgi?id=739145