project: Fix the project settings dialog to calculate the DAR consistently
The bug manifests in some cases when the w/h are linked and
you change the width to a small value and simply switch the focus
between the w/h fields. You can see slightly different values
computed for the DAR.

By making sure we set int values instead of Gst.Fraction values
to the width and height widgets we stabilize the thing.