timeline: Allow drag&drop external files on the timeline
The media library is a drag&drop source providing the URIs as text/uri-list data.
This is the same as what Nautilus provides, it must be some kind of standard.
The existing logic is changed to create the assets if they are missing. Next
time a drag-motion event occurs, with a bit of luck the assets are already
imported and the clips can be created.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=570118