timeline: Fix flickering when dragging a thin layer over a thick layer
Now there is a parameter called past_middle_when_adjacent to
__getLayerAt. If y is exactly on the layer above or below the preferred
layer, normally we'd return that layer/position.
Consider there are thin layers and thick layers.
If past_middle_when_adjacent is True, y must be far enough so that if the
layers would be switched, y will be on preferred_layer. Otherwise if the
layers are switched but y is still on the same layer, flickering takes
place.

Differential Revision: https://phabricator.freedesktop.org/D327