Stop using UIManager because it's obsolete
Refactored the management of the timeline actions. Now they are not
disabled when the timeline loses focus, because they are made available
only to the layers representation layout and the timeline toolbar. This
guarantees that the accelerators will trigger only if the focus is
the (PitiviTimeline's) layout widget. For details see
TimelineContainer._createActions.

A nice result is that the toolbar buttons don't flicker because the
actions are enabled or disabled too many times.

Fixes https://phabricator.freedesktop.org/T3416
Fixes https://phabricator.freedesktop.org/T3477

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D546