keyframes: Connect to the ControlSource value-* signals to update
Listening to the backend instead of trying to update the view ourself
is much safer, and is now possible since we implemented signals in
TimedValueSource.

It fixes the case where we split the clip since, in that case we did
not have any mean to properly know the ControlSource had changed.

Fixes T3480

Differential Revision: https://phabricator.freedesktop.org/D599