clipproperties: Fix effect configuration UI not visible
Instead of using a Gtk.Paned with the effects list at the top and the
selected effect config UI at the bottom, both in a Gtk.Box and make the
clip properties widget a Gtk.ScrolledWindow instead.

This means the effects list takes exactly the space it needs, not more
nor less.

Fixes https://phabricator.freedesktop.org/T3517

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D656