undo: Remove unuseful clean logic
A lot of clean logic is used currently only to clean
UndoableActionStacks's done_actions, populated when the action is
recorded, and undone_actions, populated when the action is undone.
It might be useful for garbage collection but currently it does not
look like a problem. This paves the way towards discarding these
objects instead of reusing them.

Differential Revision: https://phabricator.freedesktop.org/D859