viewer: Add an overlay stack only to the main viewer
Changed ViewerWidget to not be aware of the pipeline and simply display
a Gtk.Widget. The ViewerContainer now wraps the pipeline's sink widget
into an OverlayStack to provide the added functionality.

Previously the OverlayStack was always creted, even when the viewer was
used only to preview an asset.

Fixes https://phabricator.freedesktop.org/T7401

Differential Revision: https://phabricator.freedesktop.org/D945