undo: Restore the transitions details when readding a clip
When a clip which generated a transition is removed, the transition is
also removed. When the operation is undone, the transition is
recreated. The props were being lost though.

Now the action log has a new ExpandableUndoableAction class which is
able to include (merge) the actions coming right after it. We use this
for the ClipRemoved action pushed when an UriClip is removed. In our
case, first the UriClip is removed and then the transition clip(s). The
ClipRemoved is able to include now the following TransitionClipRemoved
actions and applies them when undoing *after* it has added the clip
back.

TransitionClipRemoved searches for the corresponding transition clip in
the same layer and applies the props of the video element.

Differential Revision: https://phabricator.freedesktop.org/D1095