undo: Fix undo ops on obsolete video transitions
If the user changes a transition and then later removes the clip and
then even later starts undoing and gets back to the
PropertyChangedAction corresponding to the transition change, the
action was keeping the old transition clip object and working on it
instead of the replacement.

The new UndoableAutomaticObjectAction superclass of
PropertyChangedAction keeps track of the latest object and the mapping
from old to new objects. It's notified when a replacement is found in
the TransitionClipRemoved.undo method, for now. This allows
PropertyChangedAction to work on the latest object.

PropertyChangedAction is used also for other types of objects, not only
GES.VideoTransition, but the overhead is only a dictionarly lookup when
undoing or redoing.

Differential Revision: https://phabricator.freedesktop.org/D1096