timeline: Fix traceback when opening project from the wizard
When the startup wizard is on top of the timeline and you double-click a
project in the list to be loaded, the button-release-event is triggered
for the timeline. It might be a bug in GTK, but in any case, we can
solve this very nicely by connecting the timeline's button-release-event
handler later, when the project is loaded. This required moving the
logic from new-project-created handler to new-project-loaded which seems
to work ok.

I then had to replace some connect calls with connect_after so various
initializations which depend on the timeline are done after the
timeline's initialization.

Fixes https://phabricator.freedesktop.org/T7471

Differential Revision: https://phabricator.freedesktop.org/D1106