medialibrary: Avoid switching assets when loading the project
Without this, the media library calls switchProxies when the timeline
container is not setup yet and has no ges_timeline so a backtrace is
printed. Since this call is not necessary when loading the project, a
simple check on `loaded` is enough.

Should have been part of 817889761cf5f6f59571395b2ce5d6542aa2d628

Differential Revision: https://phabricator.freedesktop.org/D1126