pipeline: Don't recover when already released
When clicking the media library Import button and the first selected
file is a video file, an AssetPipeline is started to preview it. If the
dialog is closed right away with ESC, after a few seconds
_asyncDoneNotReceivedCb is called. This one calls _recover() which puts
the pipeline in the PAUSE state which causes the preview video to show
in a new window.

Fixed by deactivating the _asyncDoneNotReceived handle.

Added a check to make sure no recovering is performed after the pipeline
has been released.

Differential Revision: https://phabricator.freedesktop.org/D1133