timeline: Remove unneeded calls
The queue_draw call in updatePosition is redundant because
updatePosition is called only in Timeline.zoomChanged and in
TimelineContainer._scrollToPixel and in both places queue_draw is called
immediately afterwards.

The removed x variable is unused.

The hiding of the ruler does not make sense since it is shown later in
the same method.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1374