timeline: Avoid redrawing the layers controls for no reason
Until now the playhead and the snapbar were being drawn by the Timeline,
and when they changed position, the layers controls had to be redrawn
along with the layers.

Moved the logic for drawing the playhead and the snapbar to a new
Gtk.Layout subclass, so when they change position, only the layers will
be redrawn.

Arguably the same performance could have been obtained by using
queue_draw_area, but the playhead and the snapbar are always drawn on
top of the layers, so it makes sense for them to be drawn by the
scrollable Layout containing the layers.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1388