timeline: Move the zoom_fitted field to Timeline
It does not make sense to keep it in the TimelineContainer since the
Timeline is zoom fitted, not the container. Moreover, this allows
getting rid of Timeline.parent which is not needed anymore.

When setting the best zoom ratio the Timeline is always scrolled to the
start. This was not the case when inserting clips.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1389