timeline: Remove Timeline._layers
Moved the __update_layers call from _removeLayer to _layerRemovedCb
because the other place where _removeLayer was called is in setProject
when all the layers were being removed, and then it does not make sense
update the remaining layers. This optimization was required because the
layer ui was already missing when __update_layers was called when
closing the project.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1430