timeline: Set the layout's size only when needed
The LayersLayout's size, along with the layers_vbox.props.width_request,
were being set when the Timeline got its allocation. This was not
optimal and caused strange vertical scroll because the vertical
adjustment is being controlled by both the LayersLayout and the layers
controls's viewport.

Now the size of the scrollable area of the LayersLayout is set in a
single place, when the layers_vbox gets its size allocation. The
layers_vbox is flagged automatically for size renegotiation in two
cases: when the heights of its Layer children change, and when the
timeline duration changes and we manually set the height_request of the
layers_vbox.

Fixes https://phabricator.freedesktop.org/T7435

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1436