elements: Get rid of more GTK warnings
Fixes a lot of "<widget> is drawn without a current allocation. This
should not happen." warnings, shown for example when adding a clip to
the timeline.

The warnings were shown because the widgets location and size were
changed when drawing. The widgets location is already being done in
various places. Now this is done also in do_map, to cover the case when
the Clip is added to the Layer.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1446