project_: Keep the assets undoable actions as the user expects
We were starting an undoable-action-log transaction when a new asset was
added to the project. This was very ugly because by the time that
transaction was closed, the user was able to make other operations on
the timeline and those were included in the initial assets-added
transaction.

This got worse with the proxy functionality since that transaction can
now take a very long time until complete, which means the undo/redo
stack was easily messed up.

Taking advantage of the fact that it does not really matter when the
assets are actually added to the project, we now record the intention of
adding them, or the intention of proxying them. Then when they are ready
to be used, the user can use them and all should be fine.

Fixes https://phabricator.freedesktop.org/T7531

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1575