mediafileprevier: Allow file discovering to be sync
In the special case we are already blocked on the
main mainloop (as we are handling the missing asset failure)
which means that we can not start another ASYNC discovery in GstDicoverer.

In the normal case, discovery is still async.

Also make the GES discovering timeout after 5 seconds only, we should not
block the UI for more than that (default is 60secs in GES).

Fixes T7645

Reviewed-by: Alex Băluț <alexandru.balut@gmail.com>
Differential Revision: https://phabricator.freedesktop.org/D1578