undo: Fix clip ungroup redo adding extra track element
When a clip is ungrouped, an empty clip is added (by GES) then a video
track element is added to it. When redoing these steps in Pitivi, GES
adds track elements itself because the clip has none, so we end up with
too many.

Added a new ClipAction undoable class to reuse the add-clip-as-is logic
between ClipAdded and ClipRemoved.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1583