render: Use a TextWidget for preset name widget
Instead of having a Gtk.Box that contains the preset widgets, use a
Gtk.Grid, allowig higher control when packing the TextWidget in code.

This will be useful so that the preset name can be properly verified
to be valid (when we will use GstEncodingTarget for the serialization
format).

Also allow using a standard GtkComboBox for the TextWidget as this is
what we need for that case.

Reviewed-by: Alex Băluț <alexandru.balut@gmail.com>
Differential Revision: https://phabricator.freedesktop.org/D1590