preset: Make use of GstEncodingTargets to handle our encoding targets
Instead of using our own grown implementation of 'RenderingPreset' use
upstream GStreamer `GstEncodingTarget` API which is more complete,
more powerful, cleaner and allows us to share the profiles with
the rest of the world, in particular with gst-transcoder which is
starting to ship profiles for most common use cases.

This also allows us to start using GstPreset to set encoders/muxer
properties which in turn makes everything much more powerful.

+ utils:ui: Return whether setting value in combo_set_value worked

Differential Revision: https://phabricator.freedesktop.org/D1596