proxy: Handle Gst.CLOCK_TIME_NONE durations
The creation_progress field was 0 because the invalid duration value was
used in a division, and it has a very large value.

Fixes https://phabricator.freedesktop.org/T7652

Also used a dict to keep the transcoded duration for each asset, instead
of using a single field holding the sum of the already transcoded
durations. This should make the computing of the ETA more precise.

Clarified  the logic in __proxyingPositionChangedCb by using more
variables.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1599