timeline: Update the layer widgets when a layer is added
When undoing a layer removal, first the layer prios were being restored,
then the layer was being added. The timeline was updating the layer
widgets only when the priority of a layer changed.

Now the layer widgets are updated also when a layer is added.

Fixes https://phabricator.freedesktop.org/T7695

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1650