timeline: Disable ungroup/group when necessary
ungroup button is now disabled when the current selection
contains only one GES_TRACK_ELEMENT or when the elements are
not in a group.

group button is disabled if the current selection elements
have the same toplevel parent.

updateActions is now called in the beginning as well, which
disables ungroup and group actions.

Closes #1699