elements: A keyframe is no longer created when moving a clip
When drag&dropping a clip, a keyframe was created if the mouse was
released on an yellow keyframe line. This happened because we were
only checking whether the click ended on the keyframe line and not
that it also started there.

Field self.__keyframe_removed is no longer needed. We were using it
to check if a keyframe was just removed, to make sure that a
double-click on a keyframe doesn't add the keyframe back after
removing it. However, we are now adding a new keyframe only if the
click occurs between 2 keyframes, so the situation described above
can't happen anymore.

Fixes T7705

Differential Revision: https://phabricator.freedesktop.org/D1682