pipeline: Fix infinite loop when a seek times out
When simple_seek was called and the operation was "successful", it was
setting the _last_position field. Then if the async operation timed out,
a recover cycle was being started and most probably the pipeline was
being set to PAUSED fine, then a new seek was attempted .. at
_last_position, then again and again.

The meaning of _last_position is clarified in this commit to be the last
position obtained successfully from the pipeline, so that such infinite
cycles cannot appear anymore.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1691