elements: Update tooltip with ts of dragged keyframe
When a Keyframe is Dragged beyond another keyframe, the tooltip updates
showing the timestamp of the hovered event.

The problem is that when a keyframe is clicked, and dragged the tooltip
is updated with the timestamp of the hovering mouse. The problem rests
with the __update_tooltip() method which updates the tooltip with the
value of the hovered event, rather than the dragged event.

To fix this, we check if the keyframe is clicked, when the
__update_tooltip() method is called, if it is clicked then we set the
tooltip to show offset of the keyframe, which does not update after it
has reached another keyframe.

Fixes T7717

Differential Revision: https://phabricator.freedesktop.org/D1695