viewer: Hide overlays when playing
Make sure only the sink widget remains visible when playing and all
the overlays are hidden. When using keyframes with the transformation
properties, the properties of the overlay are updated with a delay,
which causes the sink widget and the overlay to get out of sync. Hiding
the overlay will hide this problem.

Differential Revision: https://phabricator.freedesktop.org/D1811