elements: Lazy create the KeyFrame curves
Creating the UI representation of the keyframe is expensive (on big
projects it can take several seconds to create them - 25s for the
project provided in T7793).

We do not need to create them when creating the clip UI but only once
the clip is selected, so make sure to create them until actually
requred.

Fixes T7793

Reviewed-by: Alex Băluț <<alexandru.balut@gmail.com>>
Differential Revision: https://phabricator.freedesktop.org/D1823