previewers: Use the previewers_max_cpu setting for all previewers
And do not rely on Zoomable.app to get the info, that attribute look
quite ugly and we should try to avoid using it all around. Instead pass
it in the previewer constructor

And stop our differentiation between thumbnails and waveform generation,
it doesn't seem needed and having two options for that look too much.