mainwindow: Move the Save and Render buttons on the right side
The hamburger menu has a few export options, so this would group them
together. On the right side they are closer to the preview window, which
seems useful.

Removed the Save button icon because the word "Save" is well known, the
icon not so much, and only adds complexity.

Made the icons small.

Made the Undo/Redo buttons linked, redo button icon-only, to save some
space.

Fixes T7829

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1868