Update audio track restriction caps on project changes
Otherwise there is inconsistancy about what the timeline produce
and what the user want. For example if the user want 6 channels (5.1
audio input), what happens without this change is that the timeline
output stereo audio, and when rendering this is upmixed to 5.1, but
obviously the positionning is wrong.

Fixes T3471

Reviewed-by: Alex Băluț <<alexandru.balut@gmail.com>>
Differential Revision: https://phabricator.freedesktop.org/D1875