previewers: Simplify VideoPreviewer's queue mechanism
The mechanism between the queue and wishlist of a VideoPreviewer was
complicated, in contrast with the simple intention: to generate the
thumbnails needed by the current widgets.

Instead of using both, we now use a queue which contains from the start
exactly what is needed.

Now we display the already available thumbnails faster, without waiting
for the pipeline to start. Starting the pipeline might not even be
necessary, but that can be optimized later.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1915