previewers: Assume the thumbnail generation is accurate enough
It simplifies the logic. There is no good reason for keeping the
complicated one.

This change affects the most thumbnail placement at high zoom levels.
But in this case if the difference between the actual pixbuf position
and the expected position is so large as to be associated with a
different thumbnail, we can assume that the format of the file is not
video-editing friendly, and for these users we prefer to show all
thumbnails rather than maybe have gaps in the thumbnails.

Differential Revision: https://phabricator.freedesktop.org/D1940