render: Stop forcing format in restriction caps
This was a workaround for decoders that were allowing upstream
renegotiation while not actually supporting it, or while downstream
was fixed and encoder upstream renegotiation often leads to new
output format, thus the renegotiation was leading to NNEs

This is possible thanks to the following commit in -base.

    commit 9ce8cf89ba2d2047b4735f8a834d4659d12f91d1
    Author: Thibault Saunier <tsaunier@igalia.com>
    Date:   Sun Apr 22 10:49:29 2018 -0300

        encodebin: Also lock input caps when dynamic output is disabled

        With the way caps negotiation work in encoders, the only way to ensure
        that no downstream renegotiation is done in the encoder is to also lock
        upstream caps. Anyway with the current behavior upstream of encoders
        *require* to handle any file format so locking upstream format should
        be safe.

        https://bugzilla.gnome.org/show_bug.cgi?id=795464

Bump our GStreamer dependency to 1.14.1 as it is now required.