previewers: Fix thumbnail generation when cache db empty
When the thumbnail cache db is empty, the returned image_size is (None,
None) so thumb_width is set to None, but when there is no cache db, we
initialize it with 0. That's why thumbnail generation somehow worked for
new media files.

Fixed by using consistently 0 instead of None, and always starting the
pipeline when there is no db or the db is empty, as starting the
pipeline sets thumb_width.