previewers: Pause each previewers when pausing the GeneratorManager
To enhance performance while removing assets we are pausing
the previewer generation while deleting clips, this means that
the current previewer is always reset to None (in the Manager) and we
call `start_next_previewer` when we are done with the clip removal.
If the clip holding the generator that was being used before pausing
was not removed, it leads to that generator being restarted, but
previous code was just recreating a new pipeline from it, without
ever stoping previous pipeline, this previous pipeline was destructed
while in PAUSED... leading to the deadlock described in the bug report.

The strategy is to inform previewers that they should be
PAUSED so that they just restart when needed.

Fixes https://gitlab.gnome.org/GNOME/pitivi/issues/2227