titleeditor: Leave the focus on the timeline
When clicking a title clip, its text area was focused so you can edit
it, but the user has the impression the focus is still on the timeline
since the timeline shows the clip as selected. This gives the impression
that range selection with shift+click does not work, because the SHIFT
key-press event is not registered by the timeline, but by the textarea.

Fixes #2239