effects: Remove properties values cache
Fixes the backtraces seen in #2259.

In case of the aspectratiocrop effect, changing the aspect ratio value
leads to other props being automatically changed. The UI was
automatically updated and then the widget change triggered the same
mechanism used when the user changes a property. This consists of a new
toplevel undoable operation being created, when one was already being
recorded.

There was already a mechanism to prevent this, starting the toplevel
operation only if the value from the widget and the current property
value actually differ, but it was broken:
- It was using a cache which was not kept up to date
- The cache was holding (bool, value) instead of just value

I removed the cache as we don't see any usefulness for it.