clipproperties: Allow better control of audio properties
Audio file properties volume and mute can be controlled from
Transformations dialog of UI. This helps in easily assigning values to
various keyframes.

Most of the logic from the existing TransformationProperties class can
be found now in SourceProperties which is now the base class for the
video transformations expander and for the new audio properties
expander.

Fixes #2168

In collaboration with aleb