viewer: Improve trim preview precision and latency
The `clipTrimPreview` method is called whenever the clip trim updates,
and until now it was ignoring updates if the last performed seek was too
recent. This means the trim preview was up to 200ms old when the mouse
stopped moving.

If we want to not seek more often than 200ms, we'd have to add a timeout
in case an unwanted seek is desired, and at the end of the timeout
perform the latest wanted seek. This would be more accurate, since the
preview would be less than 200ms behind, but it would still be laggy.

To remove the lag, we shall simply seek. The `SimplePipeline.simple_seek`
method is async and already knows to skip seeks not yet performed in
case a new one is wanted.