viewer: Fix preview flicker when trimming clip
When we show the video widget of the asset backing the trimmed clip,
the first frame of the asset is shown for a brief moment until the
initial seek is performed and the desired frame is shown. This causes an
unwanted flicker.

We have to prepare the trim preview video widget and switch the project
video widget with it only when it's ready.