proxy: Use raw audio instead of opus as audio format
Opus might lead to some clipping on the raw samples and will lower
the audio Quality, the asset size is not considered to be an issue
and keeping the exact audio format is nicer in general.

Moreover the stream duration differences can lead to ugly bugs and
this makes that issue much less probable (and something would be
really wrong if it still happens with raw audio).

Closes https://gitlab.gnome.org/GNOME/pitivi/issues/2248
(it doesn't really fixes the issue for the files that are broken,
those need manual fixing sadly)