timeline: Fix click to select for tablets
Left clicking in Digitizer tablets (Wacom, Huion, etc.) is done by
tapping the pad with the stylus. However since the stylus is entirely in
the user's hand it is in constant movement due to pulse, involuntary
movements, etc. This means that after press down the cursor will always
move, which will be registered as a drag.

To fix this we set a tolerance of 3 pixels, so we can tell if the
user really wants to start dragging.

Fixes #2319