timeline: Removed current_group
Instead of keeping the current_group and the Selection object in sync,
we now use only the Selection object.

When a clip was removed from its layer, the layer was taking care of
removing it from the Selection object. This is not the case anymore,
as it simplifies the drag&drop logic. Previously as clips were dragged
from one layer to another, they were being removed from the Selection
object. Now the Selection object is stable and the logic is simpler.

Fixes #2288