Check code style with flake8
flake8 runs super fast. As it's a wrapper for pycodestyle (former pep8
tool), we use it to check for things pylint will never check, like the
number of empty lines between the methods.

We should never have to add #noqa lines to the code. Either we fix the
warning/error, or we add it to flake8's ignore list.