timeline: Fix clip handle jumping when starting drag&drop
The position passed to GES was incorrectly calculated for both left and right
handles. For the left one a 1px jump was caused by `__drag_start_x` being
between 0 and 1. For the right one, a jump of a few pixels was caused by
ignoring both `__drag_start_x` and the width of the handle, as the right edge
position must be passed to GES.

Fixes #2295