timeline: Fix drag&drop of multiple assets
When a clip cannot be placed on the timeline, we want to rollback the
previously added clips, if any. This is difficult though because GES
emits signals as if the clip which cannot be placed has been placed AND
removed. These invalid operations cannot be rolled back.

To be able to rollback, we record each clip add operation in its own
transaction. Then if it fails, we first rollback that transaction
without doing anything, then we rollback the rest so the previously
added clips are removed.

Fixes #2414