ruler.py: Refactored out timeline elements from the original ScaleRuler
Currently the ScaleRuler is unique to a Pitivi instance. The ScaleRuler displays a precise time and frame in reference to a playhead.

Since the ScaleRuler is unique to the application, an additional ruler cannot be made that is not tied to the previous ruler.

To resolve this issue, the timeline and pipeline functionality has been refactored out of ScaleRuler into a child class, TimelineScaleRuler. This will allow for additional ScaleRuler's to be created within a Pitivi instance.

Progress for Issue #2395