Clipalingment.py: Implemented easy clip alingment
Previously, there was no way to easily align clips inside
or outside of the view.

There existed alingment settings for title clips, but no easy way
to align video clips.

To resolve this issue, a new widget was made and placed in a new
"Alignment" tab. This widget allows for a user to clip in the widget
and the clip will snap to that location.

Progress towards the issue https://gitlab.gnome.org/GNOME/pitivi/issues/2279