viewer.py: Add composition guidelines to viewer
Composition guidelines are simple lines that help with
framing and position in an aesthetically pleasing way.

However, Pitivi does not provide a way to display composition
guidelines in the viewer.

To fix this, add a new overlay class (that is not attached to
a source) to draw the composition guidelines. Also, add a
menu in the viewer to select the guidelines.

https://gitlab.gnome.org/GNOME/pitivi/-/issues/2298