viewer: Allows Users to visualize and customize the safe areas for their projects
Currently, Users do not have a method to visually observe the areas of
their projects which are safe from cropping. This may lead to unwanted
cropping of their projects when they are broadcasted.

To resolve this issue, implement a new overlay class which is not
attached to a source to draw the Safe Areas, add a method to resize Safe
Areas to the project settings, implement a keyboard shortcut to toggle
the visibility of the overlay, and add documentation to the User Manual
regarding how to modify the size of a project's Safe Areas and toggle
the visibility of the overlay.

Closes #2297