Added an EditorState class
Users want to resume editing with the same interface state.

The playhead position, zoom level, scroll, and selected clips are lost.

We added an EditorState class to settings.py which is used to save and
restore these values. The EditorState is instantiated in the editor
perspective as a long-life field.

Closes GitLab issue https://gitlab.gnome.org/GNOME/pitivi/-/issues/2029