WIP: peak_meter.py: Created a VU meter
Many video editing softwares include a VU meter/peak meter that displays
the volume of audio as it is being played.

Currently Pitivi does not have a VU meter, so it lacks this convenient
method of being able to see the decibels of an audio clip change during
playback.

To fix this, two new VU meters have been added to the right of viewer,
one for left audio and one for right audio.

https://gitlab.gnome.org/GNOME/pitivi/issues/595