Viewer: Implemented a feature which allows Users to visualize and customize the safe areas for their projects
Currently, Users do not have a method to visually observe the areas of their projects which are safe from cropping.

This is an issue as users may create projects which do not comply with Safe Areas standards such as those for broadcasting. The ability to see Safe Areas needs to change in order to allow Users to visualize these areas prior to completing a project to avoid unwanted cropping of their projects.
To resolve this issue, implement a new overlay class which is not attached to a source to draw the Safe Areas, add a method to resize Safe Areas to the project settings, implement a keyboard shortcut to toggle the visibility of the overlay, and add documentation to the User Manual regarding how to modify the size of a project's Safe Areas and toggle the visibility of the overlay.

Closes Gitlab issue https://gitlab.gnome.org/GNOME/pitivi/issues/2297