timeline: Added keyframe lock feature
Added a requested feature that will lock a
keyframe to the X or Y axis when holding
Left CTRL.

This feature was added as a quality of life
upgrade to make keyframes easier to move
over the keyframe curve.

Added fields to keyframe curve class and
additional logic to determine what axis to
lock to.

Fixes #2025