test_clipproperties.py: Added unit tests for alignment editor.
Previously, there was no regression testing set up for the alignment editor.

To fix this, a test was made and placed in the ClipPropertiesTest class. This tests aligning a clip, and
undoing an action as well as re-doing the action.

Related to issue https://gitlab.gnome.org/GNOME/pitivi/issues/2279