clipproperties: Fix single-clip selection restoring
When reopening the project, the selected clip was not being properly
shown in the middle pane because the TransformationProperties expander
connects to the new-project-loaded signal after the timeline restores
the selection.

By refactoring TransformationProperties's mechanism of being notified
about the selection change, it gets properly notified of the initial
selection.

This brings TransformationProperties in line with the EffectProperties
and the other expanders, all of them being notified through
ClipProperties._selection_changed_cb now.

Fixes #2486